Expose archive API and add query structure

+++ src/schema/schema.rs
@@ -3,6 +3,6 @@ use super::Field;
 #[derive(Clone)]
 pub struct Schema {
     pub name: String,
-    pub id: u128,
+    pub id: u64,
     pub fields: Vec<Field>,
 }
